Extract image handling from the routing helpers
I've decided that helper modules in Elixir are an anti-pattern and am
now extracting things out of Alice.Router.Helpers into their own
modules.
