+++ lib/alice/router/helpers.ex
@@ -22,23 +22,9 @@ defmodule Alice.Router.Helpers do
   def reply(conn = %Conn{}, resp) when is_list(resp), do: random_reply(conn, resp)
   def reply(conn = %Conn{message: %{channel: channel}, slack: slack}, resp) do
     resp
-    |> uncache_images
+    |> Alice.Images.uncache
     |> slack_api.send_message(channel, slack)
     conn
-  end
-
-  defp uncache_images(potential_image) do
-    ~w[gif png jpg jpeg]
-    |> Enum.any?(&(potential_image |> String.downcase |> String.ends_with?(&1)))
-    |> case do
-      true -> "#{potential_image}##{random_tag}"
-      _    -> potential_image
-    end
-  end
-
-  defp random_tag do
-    "0." <> tag = to_string(:rand.uniform)
-    tag
   end
 
   defp slack_api do
